perf(ui): defer chat route rendering
